requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.2
pysafebrowsing
langdetect==1.0.9
pybind11
python-multipart==0.0.6
requests==2.31.0
pydantic==2.4.2
google-generativeai==0.3.1
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.0.1
websockets==11.0.3
aiohttp==3.8.6 
pydub